requirements.txt
opencv-python==3.4.1.15
dlib==19.13.1
numpy==1.14.5
requests==2.19.0
Pillow==5.1.0
flask==1.0.2
keras==2.2.0
tensorflow==1.8.0